+++ equalsverifier-core/src/test/java/nl/jqno/equalsverifier/integration/extra_features/nonnull/springframework/package-info.java
@@ -1,0 +1,5 @@
+/** Applies Spring's NonNullFields annotation to the package. */
+@NonNullFields
+package nl.jqno.equalsverifier.integration.extra_features.nonnull.springframework;
+
+import nl.jqno.equalsverifier.testhelpers.annotations.org.springframework.lang.NonNullFields;