+++ equalsverifier-core/src/test/java/nl/jqno/equalsverifier/testhelpers/annotations/org/jspecify/annotations/Nullable.java
@@ -1,0 +1,8 @@
+package nl.jqno.equalsverifier.testhelpers.annotations.org.jspecify.annotations;
+
+import java.lang.annotation.*;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE_USE)
+public @interface Nullable {
+}